Rename quiz components and update import paths for manage-quiz structure
