+++ app/admin/manage-quiz/_components/delete-confirmation-dialog.tsx
