+++ app/admin/manage-quiz/_components/questions-table.tsx
