+++ app/admin/manage-quiz/_components/view-dialog.tsx
